# lib/api-zod/src/generated/types/createContact.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateContactPriority } from "./createContactPriority";

export interface CreateContact {
  name: string;
  phone: string;
  relationship: string;
  priority: CreateContactPriority;
}