# lib/api-zod/src/generated/types/sosResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Guardian Companion API
 * OpenAPI spec version: 0.1.0
 */
import type { Alert } from "./alert";

export interface SosResponse {
  alert: Alert;
  trackingUrl: string;
  smsSent: boolean;
  contactsNotified: number;
}